refactor: merged isUniqueValueRaster to RasterClassifyLegend
