+++ sites/geohub/src/lib/helper/index.ts
@@ -27,7 +27,6 @@ export * from './isRasterExtension';
 export * from './localstorage';
 export * from './isStyleChanged';
 export * from './isRgbRaster';
-export * from './isUniqueValueRaster';
 export * from './isVectorIntervalExpression';
 export * from './generateHashKey';
 export * from './resolveRelativeUrl';
